requirements.txt
pytest~=8.3.3
selenium~=4.24.0
webdriver-manager~=4.0.2
Faker~=29.0.0
loguru~=0.7.2
requests~=2.32.3
allure-python-commons~=2.13.5
attrs~=24.2.0